+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # main
 - Mention the win32 device namespace in the documentation.
+- Implement `Debug` for `Settings`.
 
 # Version 0.1.7 - 2022-10-12
 - Add `write_all()` function that takes `&self`.